backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1

# ML
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# File handling
python-multipart==0.0.9
aiofiles==23.2.1
openpyxl==3.1.2

# Background jobs (optional - enable for Celery)
celery==5.3.6
redis==5.0.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3
google-auth==2.39.0
python-jose[cryptography]==3.3.0
